moved unit test about git info
