+++ common/test/test_diagnostics.py
@@ -89,40 +89,3 @@ class Diagnostics(unittest.TestCase):
         )
 
 
-class Diagnostics_FakeFS(pyfakefs_ut.TestCase):
-    """Tests using a fake filesystem.
-    """
-
-    def setUp(self):
-        self.setUpPyfakefs(allow_root_user=False)
-
-    def test_git_repo_info(self):
-
-        # not a git repo
-        self.assertEqual(diagnostics.get_git_repository_info(), None)
-
-        # simulate a git repo
-        path = pathlib.Path('.git')
-        path.mkdir()
-
-        # Branch folders and hash containing file
-        foobar = path / 'refs' / 'heads' / 'fix' / 'foobar'
-        foobar.parent.mkdir(parents=True)
-
-        with foobar.open('w') as handle:
-            handle.write('01234')
-
-        # HEAD file
-        head = path / 'HEAD'
-
-        with head.open('w') as handle:
-            handle.write('ref: refs/heads/fix/foobar')
-
-        # Test
-        self.assertEqual(
-            diagnostics.get_git_repository_info(),
-            {
-                'hash': '01234',
-                'branch': 'fix/foobar'
-            }
-        )
